phase-63: HS-63-04 the platform glue out of WebRuntime — the core is 555 lines
meeting_glue (552: start/stop, broadcast handlers, bookmarks, action
items), routing_glue (450: MIR controls, route preview, history +
artifact persistence, project association), activity (264), plugin_queue
(171), transcriber_state (144). web_runtime.py keeps boot, config apply,
presence sync, onboarding nudges, signals, run, and run_web_runtime at
555 lines (2,635 at phase open). Verbatim: 0 body lines lost.

The census proved itself the hard way: the first post-carve suite run
ABORTED with a Fatal Python error inside mlx_whisper — the device-reply
tests' web_runtime.Transcriber patch missed its moved lookup and a REAL
MLX model loaded inside a unit test. The 8 censused retargets landed (7
patch sites + the language-knob source-lock path), assertions
byte-identical, and unused imports were auto-trimmed from every new
module so no patchable name exists where nothing calls it.

Full suite 2768 passed, 17 skipped.
