+++ pm/roadmap/holdspeak/phase-63-backend-decomposition/current-phase-status.md
@@ -1,12 +1,21 @@
 # Phase 63 — Backend Decomposition
 
-**Status:** in progress (3/6). Opened 2026-06-12 on owner direction ("E,
+**Status:** in progress (4/6). Opened 2026-06-12 on owner direction ("E,
 puh-lease"): backlog row **E**, the backend twin of Phase 54. The two
 god-objects — `web_runtime.py` (2,635 lines, regrown PAST its pre-Phase-52
 size) and `meeting_session.py` (1,674) — carve into single-concern modules,
 behavior-preserving, locked by a backend density guard.
 
-**Last updated:** 2026-06-12 (**HS-63-03 done:** the feature glue is out of
+**Last updated:** 2026-06-12 (**HS-63-04 done:** the platform glue is out —
+meeting_glue (552), routing_glue (450), activity (264), plugin_queue (171),
+transcriber_state (144); **web_runtime.py is 555 lines** (boot/run/config
+only; was 2,635 at phase open). Verbatim: 0 body lines lost. The census
+proved itself the hard way: the first post-carve run ABORTED with a REAL
+MLX model loading inside a unit test because the web_runtime.Transcriber
+patch missed its moved lookup — the 8 censused retargets landed (7 patch
+sites + 1 source-lock path), assertions byte-identical; unused imports
+auto-trimmed from every module so no patchable name exists where nothing
+calls it. Suite **2768 passed, 17 skipped**. Prior: **HS-63-03 done:** the feature glue is out of
 WebRuntime — dictation_capture (413: transcribe-and-type, hotkey, tmux,
 voice commands), wake_glue (~360, the HS-60 section verbatim), device_glue
 (~315); web_runtime.py is **1,763** (was 2,635). The verbatim proof is again
@@ -82,12 +91,12 @@ guard locks the shape.
 | HS-63-01 | The meeting models | done | none |
 | HS-63-02 | MeetingSession mixins | done | HS-63-01 |
 | HS-63-03 | WebRuntime mixins: the feature glue | done | none |
-| HS-63-04 | WebRuntime mixins: the platform glue + thin core | backlog | HS-63-03 |
+| HS-63-04 | WebRuntime mixins: the platform glue + thin core | done | HS-63-03 |
 | HS-63-05 | The backend density guard + docs | backlog | HS-63-01..04 |
 | HS-63-06 | Closeout: the live boot proof + final-summary + PR | backlog | HS-63-01..05 |
 
 ## Where we are
 
-Both feature carves are done (meeting side largest 795; runtime at 1,763).
-Next is **HS-63-04 — the platform glue + thin core** (meeting glue, MIR/
-routing, the plugin queue, activity, transcriber state).
+Both god-objects are carved: web_runtime 2,635 → 555; meeting_session
+1,674 → 795 core; thirteen single-concern modules, all ≤600. Next is
+**HS-63-05 — the backend density guard + docs**.
