+++ tests/unit/test_language_knob.py
@@ -179,7 +179,9 @@ def test_all_construction_sites_thread_the_language() -> None:
 
     repo = Path(__file__).resolve().parents[2]
     sites = {
-        "holdspeak/web_runtime.py": 'language=getattr(self.config.model, "language", "auto")',
+        # HS-63-04: the runtime's Transcriber construction moved with its
+        # method (_ensure_transcriber_loaded) into the transcriber_state mixin.
+        "holdspeak/runtime/transcriber_state.py": 'language=getattr(self.config.model, "language", "auto")',
         "holdspeak/main.py": 'language=getattr(config.model, "language", "auto")',
         "holdspeak/web/routes/meeting_import.py": 'language=getattr(config.model, "language", "auto")',
         "holdspeak/commands/import_recording.py": 'language=getattr(config.model, "language", "auto")',
